Move test_cov and test_mvn_log_pdf into the test suite
Removes the inline test functions and __main__ block from integration.py
and adds them as proper pytest tests in tests/algorithms/test_integration.py.
test_cov is parameterized over ddof=0 and ddof=1.
