+++ src/laue_dials/algorithms/integration.py
@@ -334,34 +334,6 @@ def cov(m, aweights=None, return_mean=False, ddof=0):
     return S
 
 
-def test_cov(ddof=0):
-    d = 10
-    n = 1_000
-    b = 5
-
-    from scipy.stats import multivariate_normal
-
-    loc = np.random.rand(d)
-    L = np.tril(np.random.rand(d * d).reshape((d, d)))
-    S = L.T @ L
-
-    m = multivariate_normal.rvs(loc, S, size=(b, n))
-    if b == 1:
-        m = m[None, ...]
-
-    expected = np.stack([np.cov(i.T, ddof=ddof) for i in m])
-    result = cov(m, ddof=ddof)
-    assert np.allclose(expected, result)
-
-    aweights = np.random.rand(b * n).reshape((b, n))
-
-    expected = np.stack(
-        [np.cov(i.T, aweights=a, ddof=ddof) for i, a in zip(m, aweights)]
-    )
-    result = cov(m, aweights=aweights[..., None], ddof=ddof)
-    assert np.allclose(expected, result)
-
-
 def mvn_log_pdf(x, loc, scale, return_zscore=False):
     """a batched version of scipy.stats.multivariate_normal.log_pdf"""
     d = loc.shape[-1]
@@ -376,28 +348,3 @@ def mvn_log_pdf(x, loc, scale, return_zscore=False):
     if return_zscore:
         return log_p, zscore
     return log_p
-
-
-def test_mvn_log_pdf():
-    d = 10
-    b = 5
-    m = 50
-    X = np.random.random(b * m * d).reshape((b, m, d))  # test points
-
-    from scipy.stats import multivariate_normal
-
-    loc = np.random.rand(b * d).reshape(b, d)
-    L = np.tril(np.random.rand(b * d * d).reshape((b, d, d))) + np.eye(d)
-    S = L.swapaxes(-1, -2) @ L
-
-    expected = np.stack(
-        [multivariate_normal.logpdf(i, j, k) for i, j, k in zip(X, loc, S)]
-    )
-    result = mvn_log_pdf(X, loc, S)
-    assert np.allclose(expected, result)
-
-
-if __name__ == "__main__":
-    test_mvn_log_pdf()
-    test_cov(ddof=0)
-    test_cov(ddof=1)
