+++ tests/algorithms/test_integration.py
@@ -1,33 +1,52 @@
 import numpy as np
 import pytest
 
-from laue_dials.algorithms.integration import Profile, SegmentedImage
+from laue_dials.algorithms.integration import cov, mvn_log_pdf
 
 
-@pytest.fixture
-def example_profile_data():
-    x = np.linspace(1, 1000, num=1000)
-    y = np.linspace(1, 1000, num=1000)
-    counts = np.random.randint(100, 1000, size=len(x))
-    return x, y, counts
+@pytest.mark.parametrize("ddof", [0, 1])
+def test_cov(ddof):
+    d = 10
+    n = 1_000
+    b = 5
+
+    from scipy.stats import multivariate_normal
+
+    loc = np.random.rand(d)
+    L = np.tril(np.random.rand(d * d).reshape((d, d)))
+    S = L.T @ L
+
+    m = multivariate_normal.rvs(loc, S, size=(b, n))
+    if b == 1:
+        m = m[None, ...]
+
+    expected = np.stack([np.cov(i.T, ddof=ddof) for i in m])
+    result = cov(m, ddof=ddof)
+    assert np.allclose(expected, result)
+
+    aweights = np.random.rand(b * n).reshape((b, n))
+
+    expected = np.stack(
+        [np.cov(i.T, aweights=a, ddof=ddof) for i, a in zip(m, aweights)]
+    )
+    result = cov(m, aweights=aweights[..., None], ddof=ddof)
+    assert np.allclose(expected, result)
 
 
-def test_profile_init(example_profile_data):
-    x, y, counts = example_profile_data
-    profile = Profile(x, y, counts)
-    assert profile is not None
-    # Add more assertions based on your expectations
+def test_mvn_log_pdf():
+    d = 10
+    b = 5
+    m = 50
+    X = np.random.random(b * m * d).reshape((b, m, d))
 
+    from scipy.stats import multivariate_normal
 
-@pytest.fixture
-def example_segmented_image_data():
-    # Replace this with actual data for creating a SegmentedImage instance
-    pixels = np.array([[0, 0, 1], [1, 1, 0], [0, 1, 0]])
-    centroids = np.array([[0, 0], [1, 1], [2, 2]])
-    return pixels, centroids
+    loc = np.random.rand(b * d).reshape(b, d)
+    L = np.tril(np.random.rand(b * d * d).reshape((b, d, d))) + np.eye(d)
+    S = L.swapaxes(-1, -2) @ L
 
-
-def test_segmented_image_init(example_segmented_image_data):
-    pixels, centroids = example_segmented_image_data
-    segmented_image = SegmentedImage(pixels, centroids)
-    assert segmented_image is not None
+    expected = np.stack(
+        [multivariate_normal.logpdf(i, j, k) for i, j, k in zip(X, loc, S)]
+    )
+    result = mvn_log_pdf(X, loc, S)
+    assert np.allclose(expected, result)
